Shop System in Progress
More functionalities are applied to the shop. You can see name, cost and description when hovering over an item when shopping. If the player press an item a messege will pop up asking if the player is sure of the purches or telling the player has not enough money for the item.